Move import handler to PackageImportDialog
