+++ i18n/pt.po
@@ -1,7 +1,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: i18next-conv\n"
-"POT-Creation-Date: 2020-10-20T09:26:11.585Z\n"
+"POT-Creation-Date: 2020-10-20T10:33:16.288Z\n"
 "PO-Revision-Date: 2020-07-10T06:53:30.625Z\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
@@ -524,6 +524,12 @@ msgstr ""
 msgid "[Pull Request by {{name}}] {{subject}}"
 msgstr ""
 
+msgid "Importing package {{name}}"
+msgstr ""
+
+msgid "Couldn't load package"
+msgstr ""
+
 msgid "Packages Import"
 msgstr ""
 
@@ -572,12 +578,6 @@ msgid "Package published to store in a new branch"
 msgstr ""
 
 msgid "Couldn't create new branch on store"
-msgstr ""
-
-msgid "Importing package {{name}}"
-msgstr ""
-
-msgid "Couldn't load package"
 msgstr ""
 
 msgid "Version"
